requirements.txt
pyyaml
psutil
py-cpuinfo
pandas